+++ components/vc-select/Selector/SingleSelector.tsx
@@ -3,8 +3,8 @@ import Input from './Input';
 import type { InnerSelectorProps } from './interface';
 import { Fragment, computed, defineComponent, ref, watch } from 'vue';
 import PropTypes from '../../_util/vue-types';
-import { useInjectTreeSelectContext } from '../../vc-tree-select/Context';
 import type { VueNode } from '../../_util/type';
+import useInjectLegacySelectContext from '../../vc-tree-select/LegacyContext';
 
 interface SelectorProps extends InnerSelectorProps {
   inputElement: VueNode;
@@ -50,7 +50,7 @@ const SingleSelector = defineComponent<SelectorProps>({
       }
       return inputValue;
     });
-    const treeSelectContext = useInjectTreeSelectContext();
+    const legacyTreeSelectContext = useInjectLegacySelectContext();
     watch(
       [combobox, () => props.activeValue],
       () => {
@@ -108,13 +108,16 @@ const SingleSelector = defineComponent<SelectorProps>({
       const item = values[0];
       let titleNode = null;
       // custom tree-select title by slot
-      if (item && treeSelectContext.value.slots) {
+
+      if (item && legacyTreeSelectContext.customSlots) {
+        const key = item.key ?? item.value;
+        const originData = legacyTreeSelectContext.keyEntities[key]?.node || {};
         titleNode =
-          treeSelectContext.value.slots[item?.option?.data?.slots?.title] ||
-          treeSelectContext.value.slots.title ||
+          legacyTreeSelectContext.customSlots[originData.slots?.title] ||
+          legacyTreeSelectContext.customSlots.title ||
           item.label;
         if (typeof titleNode === 'function') {
-          titleNode = titleNode(item.option?.data || {});
+          titleNode = titleNode(originData);
         }
         //  else if (treeSelectContext.value.slots.titleRender) {
         //   // 因历史 title 是覆盖逻辑，新增 titleRender，所有的 title 都走一遍 titleRender
@@ -155,7 +158,7 @@ const SingleSelector = defineComponent<SelectorProps>({
           {/* Display value */}
           {!combobox.value && item && !hasTextInput.value && (
             <span class={`${prefixCls}-selection-item`} title={title.value}>
-              <Fragment key={item.key || item.value}>{titleNode}</Fragment>
+              <Fragment key={item.key ?? item.value}>{titleNode}</Fragment>
             </span>
           )}
