+++ components/vc-tree-select/index.tsx
@@ -1,9 +1,8 @@
-// base rc-tree-select@4.6.1
-import TreeSelect from './TreeSelect';
+// base rc-tree-select@5.0.0-alpha.4
+import type { TreeSelectProps } from './TreeSelect';
+import TreeSelect, { treeSelectProps } from './TreeSelect';
 import TreeNode from './TreeNode';
 import { SHOW_ALL, SHOW_CHILD, SHOW_PARENT } from './utils/strategyUtil';
-import type { TreeSelectProps } from './props';
-import { treeSelectProps } from './props';
 
 export { TreeNode, SHOW_ALL, SHOW_CHILD, SHOW_PARENT, treeSelectProps };
 export type { TreeSelectProps };
