+++ components/vc-tree-select/utils/legacyUtil.tsx
@@ -1,21 +1,15 @@
 import { filterEmpty } from '../../_util/props-util';
 import { camelize } from 'vue';
 import { warning } from '../../vc-util/warning';
-import type {
-  DataNode,
-  LegacyDataNode,
-  ChangeEventExtra,
-  InternalDataEntity,
-  RawValueType,
-  LegacyCheckedNode,
-} from '../interface';
+import type { DataNode, ChangeEventExtra, RawValueType, LegacyCheckedNode } from '../interface';
 import TreeNode from '../TreeNode';
 import type { VueNode } from '../../_util/type';
+import type { DefaultOptionType, FieldNames } from '../TreeSelect';
 
 function isTreeSelectNode(node: any) {
   return node && node.type && (node.type as any).isTreeSelectNode;
 }
-export function convertChildrenToData(rootNodes: VueNode): DataNode[] {
+export function convertChildrenToData(rootNodes: VueNode[]): DataNode[] {
   function dig(treeNodes: any[] = []): DataNode[] {
     return filterEmpty(treeNodes).map(treeNode => {
       // Filter invalidate node
@@ -66,10 +60,10 @@ export function convertChildrenToData(rootNodes: VueNode): DataNode[] {
   return dig(rootNodes as any[]);
 }
 
-export function fillLegacyProps(dataNode: DataNode): LegacyDataNode {
+export function fillLegacyProps(dataNode: DataNode): any {
   // Skip if not dataNode exist
   if (!dataNode) {
-    return dataNode as LegacyDataNode;
+    return dataNode;
   }
 
   const cloneNode = { ...dataNode };
@@ -79,37 +73,43 @@ export function fillLegacyProps(dataNode: DataNode): LegacyDataNode {
       get() {
         warning(
           false,
-          'New `rc-tree-select` not support return node instance as argument anymore. Please consider to remove `props` access.',
+          'New `vc-tree-select` not support return node instance as argument anymore. Please consider to remove `props` access.',
         );
         return cloneNode;
       },
     });
   }
 
-  return cloneNode as LegacyDataNode;
+  return cloneNode;
 }
 
 export function fillAdditionalInfo(
   extra: ChangeEventExtra,
   triggerValue: RawValueType,
   checkedValues: RawValueType[],
-  treeData: InternalDataEntity[],
+  treeData: DefaultOptionType[],
   showPosition: boolean,
+  fieldNames: FieldNames,
 ) {
   let triggerNode = null;
   let nodeList: LegacyCheckedNode[] = null;
 
   function generateMap() {
-    function dig(list: InternalDataEntity[], level = '0', parentIncluded = false) {
+    function dig(list: DefaultOptionType[], level = '0', parentIncluded = false) {
       return list
-        .map((dataNode, index) => {
+        .map((option, index) => {
           const pos = `${level}-${index}`;
-          const included = checkedValues.includes(dataNode.value);
-          const children = dig(dataNode.children || [], pos, included);
-          const node = <TreeNode {...dataNode}>{children.map(child => child.node)}</TreeNode>;
+          const value = option[fieldNames.value];
+          const included = checkedValues.includes(value);
+          const children = dig(option[fieldNames.children] || [], pos, included);
+          const node = (
+            <TreeNode {...(option as Required<DefaultOptionType>)}>
+              {children.map(child => child.node)}
+            </TreeNode>
+          );
 
           // Link with trigger node
-          if (triggerValue === dataNode.value) {
+          if (triggerValue === value) {
             triggerNode = node;
           }
