+++ components/vc-tree-select/utils/strategyUtil.ts
@@ -1,5 +1,6 @@
 import type { DataEntity } from '../../vc-tree/interface';
-import type { RawValueType, Key, DataNode } from '../interface';
+import type { InternalFieldName } from '../TreeSelect';
+import type { RawValueType, Key } from '../interface';
 import { isCheckDisabled } from './valueUtil';
 
 export const SHOW_ALL = 'SHOW_ALL';
@@ -8,22 +9,23 @@ export const SHOW_CHILD = 'SHOW_CHILD';
 
 export type CheckedStrategy = typeof SHOW_ALL | typeof SHOW_PARENT | typeof SHOW_CHILD;
 
-export function formatStrategyKeys(
-  keys: Key[],
+export function formatStrategyValues(
+  values: Key[],
   strategy: CheckedStrategy,
   keyEntities: Record<Key, DataEntity>,
+  fieldNames: InternalFieldName,
 ): RawValueType[] {
-  const keySet = new Set(keys);
+  const valueSet = new Set(values);
 
   if (strategy === SHOW_CHILD) {
-    return keys.filter(key => {
+    return values.filter(key => {
       const entity = keyEntities[key];
 
       if (
         entity &&
         entity.children &&
         entity.children.every(
-          ({ node }) => isCheckDisabled(node) || keySet.has((node as DataNode).key),
+          ({ node }) => isCheckDisabled(node) || valueSet.has(node[fieldNames.value]),
         )
       ) {
         return false;
@@ -32,15 +34,14 @@ export function formatStrategyKeys(
     });
   }
   if (strategy === SHOW_PARENT) {
-    return keys.filter(key => {
+    return values.filter(key => {
       const entity = keyEntities[key];
       const parent = entity ? entity.parent : null;
-
-      if (parent && !isCheckDisabled(parent.node) && keySet.has((parent.node as DataNode).key)) {
+      if (parent && !isCheckDisabled(parent.node) && valueSet.has(parent.key)) {
         return false;
       }
       return true;
     });
   }
-  return keys;
+  return values;
 }
