+++ components/vc-tree-select/utils/warningPropsUtil.ts
@@ -1,7 +1,8 @@
 import { warning } from '../../vc-util/warning';
+import type { TreeSelectProps } from '../TreeSelect';
 import { toArray } from './valueUtil';
 
-function warningProps(props: any) {
+function warningProps(props: TreeSelectProps & { searchPlaceholder?: string }) {
   const { searchPlaceholder, treeCheckStrictly, treeCheckable, labelInValue, value, multiple } =
     props;
