+++ components/vc-tree-select1/OptionList.tsx
@@ -1,16 +1,14 @@
-import type { TreeDataNode, Key } from './interface';
+import type { DataNode, TreeDataNode, Key } from './interface';
+import { useInjectTreeSelectContext } from './Context';
 import type { RefOptionListProps } from '../vc-select/OptionList';
 import type { ScrollTo } from '../vc-virtual-list/List';
 import { computed, defineComponent, nextTick, ref, shallowRef, watch } from 'vue';
+import { optionListProps } from './props';
 import useMemo from '../_util/hooks/useMemo';
 import type { EventDataNode } from '../tree';
 import KeyCode from '../_util/KeyCode';
 import Tree from '../vc-tree/Tree';
 import type { TreeProps } from '../vc-tree/props';
-import { getAllKeys, isCheckDisabled } from './utils/valueUtil';
-import { useBaseProps } from '../vc-select';
-import useInjectLegacySelectContext from './LegacyContext';
-import useInjectSelectContext from './TreeSelectContext';
 
 const HIDDEN_STYLE = {
   width: 0,
@@ -34,36 +32,44 @@ type ReviseRefOptionListProps = Omit<RefOptionListProps, 'scrollTo'> & { scrollT
 export default defineComponent({
   name: 'OptionList',
   inheritAttrs: false,
+  props: optionListProps<DataNode>(),
   slots: ['notFoundContent', 'menuItemSelectedIcon'],
-  setup(_, { slots, expose }) {
-    const baseProps = useBaseProps();
-    const legacyContext = useInjectLegacySelectContext();
-    const context = useInjectSelectContext();
+  setup(props, { slots, expose }) {
+    const context = useInjectTreeSelectContext();
+
     const treeRef = ref();
-    const memoTreeData = useMemo(
-      () => context.treeData,
-      [() => baseProps.open, () => context.treeData],
+    const memoOptions = useMemo(
+      () => props.options,
+      [() => props.open, () => props.options],
       next => next[0],
     );
 
+    const valueKeys = computed(() => {
+      const { checkedKeys, getEntityByValue } = context.value;
+      return checkedKeys.map(val => {
+        const entity = getEntityByValue(val);
+        return entity ? entity.key : null;
+      });
+    });
+
     const mergedCheckedKeys = computed(() => {
-      const { checkable, halfCheckedKeys, checkedKeys } = legacyContext;
+      const { checkable, halfCheckedKeys } = context.value;
       if (!checkable) {
         return null;
       }
 
       return {
-        checked: checkedKeys,
+        checked: valueKeys.value,
         halfChecked: halfCheckedKeys,
       };
     });
 
     watch(
-      () => baseProps.open,
+      () => props.open,
       () => {
         nextTick(() => {
-          if (baseProps.open && !baseProps.multiple && legacyContext.checkedKeys.length) {
-            treeRef.value?.scrollTo({ key: legacyContext.checkedKeys[0] });
+          if (props.open && !props.multiple && valueKeys.value.length) {
+            treeRef.value?.scrollTo({ key: valueKeys.value[0] });
           }
         });
       },
@@ -71,25 +77,25 @@ export default defineComponent({
     );
 
     // ========================== Search ==========================
-    const lowerSearchValue = computed(() => String(baseProps.searchValue).toLowerCase());
+    const lowerSearchValue = computed(() => String(props.searchValue).toLowerCase());
     const filterTreeNode = (treeNode: EventDataNode) => {
       if (!lowerSearchValue.value) {
         return false;
       }
-      return String(treeNode[legacyContext.treeNodeFilterProp])
+      return String(treeNode[context.value.treeNodeFilterProp])
         .toLowerCase()
         .includes(lowerSearchValue.value);
     };
 
     // =========================== Keys ===========================
-    const expandedKeys = shallowRef<Key[]>(legacyContext.treeDefaultExpandedKeys);
+    const expandedKeys = shallowRef<Key[]>(context.value.treeDefaultExpandedKeys);
     const searchExpandedKeys = shallowRef<Key[]>(null);
 
     watch(
-      () => baseProps.searchValue,
+      () => props.searchValue,
       () => {
-        if (baseProps.searchValue) {
-          searchExpandedKeys.value = getAllKeys(context.treeData, context.fieldNames);
+        if (props.searchValue) {
+          searchExpandedKeys.value = props.flattenOptions.map(o => o.key);
         }
       },
       {
@@ -97,17 +103,17 @@ export default defineComponent({
       },
     );
     const mergedExpandedKeys = computed(() => {
-      if (legacyContext.treeExpandedKeys) {
-        return [...legacyContext.treeExpandedKeys];
+      if (context.value.treeExpandedKeys) {
+        return [...context.value.treeExpandedKeys];
       }
-      return baseProps.searchValue ? searchExpandedKeys.value : expandedKeys.value;
+      return props.searchValue ? searchExpandedKeys.value : expandedKeys.value;
     });
 
     const onInternalExpand = (keys: Key[]) => {
       expandedKeys.value = keys;
       searchExpandedKeys.value = keys;
 
-      legacyContext.onTreeExpand?.(keys);
+      context.value.onTreeExpand?.(keys);
     };
 
     // ========================== Events ==========================
@@ -115,23 +121,23 @@ export default defineComponent({
       event.preventDefault();
     };
 
-    const onInternalSelect = (_: Key[], { node }: TreeEventInfo) => {
-      const { checkable, checkedKeys } = legacyContext;
-      if (checkable && isCheckDisabled(node)) {
-        return;
+    const onInternalSelect = (_: Key[], { node: { key } }: TreeEventInfo) => {
+      const { getEntityByKey, checkable, checkedKeys } = context.value;
+      const entity = getEntityByKey(key, checkable ? 'checkbox' : 'select');
+      if (entity !== null) {
+        props.onSelect?.(entity.data.value, {
+          selected: !checkedKeys.includes(entity.data.value),
+        });
       }
-      context.onSelect?.(node.key, {
-        selected: !checkedKeys.includes(node.key),
-      });
 
-      if (!baseProps.multiple) {
-        baseProps.toggleOpen?.(false);
+      if (!props.multiple) {
+        props.onToggleOpen?.(false);
       }
     };
 
     // ========================= Keyboard =========================
     const activeKey = ref<Key>(null);
-    const activeEntity = computed(() => legacyContext.keyEntities[activeKey.value]);
+    const activeEntity = computed(() => context.value.getEntityByKey(activeKey.value));
 
     const setActiveKey = (key: Key) => {
       activeKey.value = key;
@@ -151,11 +157,11 @@ export default defineComponent({
 
           // >>> Select item
           case KeyCode.ENTER: {
-            const { selectable, value } = activeEntity.value?.node || {};
+            const { selectable, value } = activeEntity.value?.data.node || {};
             if (selectable !== false) {
               onInternalSelect(null, {
                 node: { key: activeKey.value },
-                selected: !legacyContext.checkedKeys.includes(value),
+                selected: !context.value.checkedKeys.includes(value),
               });
             }
             break;
@@ -163,7 +169,7 @@ export default defineComponent({
 
           // >>> Close
           case KeyCode.ESC: {
-            baseProps.toggleOpen(false);
+            props.onToggleOpen(false);
           }
         }
       },
@@ -173,12 +179,15 @@ export default defineComponent({
     return () => {
       const {
         prefixCls,
+        height,
+        itemHeight,
+        virtual,
         multiple,
         searchValue,
         open,
         notFoundContent = slots.notFoundContent?.(),
-      } = baseProps;
-      const { listHeight, listItemHeight, virtual } = context;
+        onMouseenter,
+      } = props;
       const {
         checkable,
         treeDefaultExpandAll,
@@ -190,10 +199,9 @@ export default defineComponent({
         treeLoadedKeys,
         treeMotion,
         onTreeLoad,
-        checkedKeys,
-      } = legacyContext;
+      } = context.value;
       // ========================== Render ==========================
-      if (memoTreeData.value.length === 0) {
+      if (memoOptions.value.length === 0) {
         return (
           <div role="listbox" class={`${prefixCls}-empty`} onMousedown={onListMouseDown}>
             {notFoundContent}
@@ -209,10 +217,10 @@ export default defineComponent({
         treeProps.expandedKeys = mergedExpandedKeys.value;
       }
       return (
-        <div onMousedown={onListMouseDown}>
+        <div onMousedown={onListMouseDown} onMouseenter={onMouseenter}>
           {activeEntity.value && open && (
             <span style={HIDDEN_STYLE} aria-live="assertive">
-              {activeEntity.value.node.value}
+              {activeEntity.value.data.value}
             </span>
           )}
 
@@ -220,9 +228,9 @@ export default defineComponent({
             ref={treeRef}
             focusable={false}
             prefixCls={`${prefixCls}-tree`}
-            treeData={memoTreeData.value as TreeDataNode[]}
-            height={listHeight}
-            itemHeight={listItemHeight}
+            treeData={memoOptions.value as TreeDataNode[]}
+            height={height}
+            itemHeight={itemHeight}
             virtual={virtual}
             multiple={multiple}
             icon={treeIcon}
@@ -235,7 +243,7 @@ export default defineComponent({
             checkable={checkable}
             checkStrictly
             checkedKeys={mergedCheckedKeys.value}
-            selectedKeys={!checkable ? checkedKeys : []}
+            selectedKeys={!checkable ? valueKeys.value : []}
             defaultExpandAll={treeDefaultExpandAll}
             {...treeProps}
             // Proxy event out
@@ -245,7 +253,7 @@ export default defineComponent({
             onExpand={onInternalExpand}
             onLoad={onTreeLoad}
             filterTreeNode={filterTreeNode}
-            v-slots={{ ...slots, checkable: legacyContext.customCheckable }}
+            v-slots={{ ...slots, checkable: context.value.customCheckable }}
           />
         </div>
       );
