+++ components/vc-tree-select1/interface.ts
