+++ components/vc-tree-select1/utils/legacyUtil.tsx
@@ -1,15 +1,21 @@
 import { filterEmpty } from '../../_util/props-util';
 import { camelize } from 'vue';
 import { warning } from '../../vc-util/warning';
-import type { DataNode, ChangeEventExtra, RawValueType, LegacyCheckedNode } from '../interface';
+import type {
+  DataNode,
+  LegacyDataNode,
+  ChangeEventExtra,
+  InternalDataEntity,
+  RawValueType,
+  LegacyCheckedNode,
+} from '../interface';
 import TreeNode from '../TreeNode';
 import type { VueNode } from '../../_util/type';
-import type { DefaultOptionType, FieldNames } from '../TreeSelect';
 
 function isTreeSelectNode(node: any) {
   return node && node.type && (node.type as any).isTreeSelectNode;
 }
-export function convertChildrenToData(rootNodes: VueNode[]): DataNode[] {
+export function convertChildrenToData(rootNodes: VueNode): DataNode[] {
   function dig(treeNodes: any[] = []): DataNode[] {
     return filterEmpty(treeNodes).map(treeNode => {
       // Filter invalidate node
@@ -60,10 +66,10 @@ export function convertChildrenToData(rootNodes: VueNode[]): DataNode[] {
   return dig(rootNodes as any[]);
 }
 
-export function fillLegacyProps(dataNode: DataNode): any {
+export function fillLegacyProps(dataNode: DataNode): LegacyDataNode {
   // Skip if not dataNode exist
   if (!dataNode) {
-    return dataNode;
+    return dataNode as LegacyDataNode;
   }
 
   const cloneNode = { ...dataNode };
@@ -73,43 +79,37 @@ export function fillLegacyProps(dataNode: DataNode): any {
       get() {
         warning(
           false,
-          'New `vc-tree-select` not support return node instance as argument anymore. Please consider to remove `props` access.',
+          'New `rc-tree-select` not support return node instance as argument anymore. Please consider to remove `props` access.',
         );
         return cloneNode;
       },
     });
   }
 
-  return cloneNode;
+  return cloneNode as LegacyDataNode;
 }
 
 export function fillAdditionalInfo(
   extra: ChangeEventExtra,
   triggerValue: RawValueType,
   checkedValues: RawValueType[],
-  treeData: DefaultOptionType[],
+  treeData: InternalDataEntity[],
   showPosition: boolean,
-  fieldNames: FieldNames,
 ) {
   let triggerNode = null;
   let nodeList: LegacyCheckedNode[] = null;
 
   function generateMap() {
-    function dig(list: DefaultOptionType[], level = '0', parentIncluded = false) {
+    function dig(list: InternalDataEntity[], level = '0', parentIncluded = false) {
       return list
-        .map((option, index) => {
+        .map((dataNode, index) => {
           const pos = `${level}-${index}`;
-          const value = option[fieldNames.value];
-          const included = checkedValues.includes(value);
-          const children = dig(option[fieldNames.children] || [], pos, included);
-          const node = (
-            <TreeNode {...(option as Required<DefaultOptionType>)}>
-              {children.map(child => child.node)}
-            </TreeNode>
-          );
+          const included = checkedValues.includes(dataNode.value);
+          const children = dig(dataNode.children || [], pos, included);
+          const node = <TreeNode {...dataNode}>{children.map(child => child.node)}</TreeNode>;
 
           // Link with trigger node
-          if (triggerValue === value) {
+          if (triggerValue === dataNode.value) {
             triggerNode = node;
           }
