+++ components/vc-tree-select1/utils/strategyUtil.ts
@@ -1,6 +1,5 @@
 import type { DataEntity } from '../../vc-tree/interface';
-import type { InternalFieldName } from '../TreeSelect';
-import type { RawValueType, Key } from '../interface';
+import type { RawValueType, Key, DataNode } from '../interface';
 import { isCheckDisabled } from './valueUtil';
 
 export const SHOW_ALL = 'SHOW_ALL';
@@ -9,23 +8,22 @@ export const SHOW_CHILD = 'SHOW_CHILD';
 
 export type CheckedStrategy = typeof SHOW_ALL | typeof SHOW_PARENT | typeof SHOW_CHILD;
 
-export function formatStrategyValues(
-  values: Key[],
+export function formatStrategyKeys(
+  keys: Key[],
   strategy: CheckedStrategy,
   keyEntities: Record<Key, DataEntity>,
-  fieldNames: InternalFieldName,
 ): RawValueType[] {
-  const valueSet = new Set(values);
+  const keySet = new Set(keys);
 
   if (strategy === SHOW_CHILD) {
-    return values.filter(key => {
+    return keys.filter(key => {
       const entity = keyEntities[key];
 
       if (
         entity &&
         entity.children &&
         entity.children.every(
-          ({ node }) => isCheckDisabled(node) || valueSet.has(node[fieldNames.value]),
+          ({ node }) => isCheckDisabled(node) || keySet.has((node as DataNode).key),
         )
       ) {
         return false;
@@ -34,15 +32,15 @@ export function formatStrategyValues(
     });
   }
   if (strategy === SHOW_PARENT) {
-    return values.filter(key => {
+    return keys.filter(key => {
       const entity = keyEntities[key];
       const parent = entity ? entity.parent : null;
 
-      if (parent && !isCheckDisabled(parent.node) && valueSet.has(parent.node.key)) {
+      if (parent && !isCheckDisabled(parent.node) && keySet.has((parent.node as DataNode).key)) {
         return false;
       }
       return true;
     });
   }
-  return values;
+  return keys;
 }
